use double for hdop

+++ src/navigation/11_Gps.ino
@@ -67,7 +67,7 @@ class Gps
   int satellites;
   unsigned fix_age, time, date;
   float speed, course;
-  unsigned hdop;
+  double hdop;
 
   private:
